# Adding `--cuda_allow_inline_execution` flag.