This allows command buffers with the
IREE_HAL_COMMAND_BUFFER_MODE_ALLOW_INLINE_EXECUTION flag set to execute
directly against a CUDA stream to lower latency.

Disabled by default because this is slower in almost all cases where the
model actually uses the GPU to do work. simple_embedding_test (a single
dispatch) for example goes 39.47us -> 39.32us with this enabled, while
mobilessd (GPU-bound) goes 2.56s -> 2.69s.